+++ src/components/Explorer/Base/Result/ChartHelpers.jsx
