+++ src/components/Explorer/Base/Result/MatchingRuns/SerratusApiCalls.jsx
@@ -2,11 +2,6 @@ import axios from 'axios'
 
 const baseUrl = process.env.REACT_APP_API_URL;
 
-export const fetchSraRun = async (searchType, runId) => {
-    const response = await axios.get(`${baseUrl}/summary/${searchType}/run=${runId}`);
-    return response.data;
-}
-
 export const getMatchesDownloadUrl = (searchType, searchLevel, searchLevelValue, identityLims, scoreLims) => {
     var [identityMin, identityMax] = identityLims;
     var [scoreMin, scoreMax] = scoreLims;
